prod urls added

+++ Backend/src/utils/cronAutoDelete.js
@@ -1,8 +1,13 @@
 const cron = require("node-cron");
 const Notes = require("../schema/noteSchema");
+const { default: mongoose } = require("mongoose");
 
 cron.schedule("* * * * *", async () => {
   const now = new Date();
+  if (mongoose.connection.readyState !== 1) {
+    console.warn("MongoDB not connected yet. Skipping this auto-delete cycle");
+    return;
+  }
   try {
     const result = await Notes.deleteMany({ deleteAt: { $lte: now } });
     if (result.deletedCount > 0) {
